// src/Types/car.tsx
export interface NewCarType{
    model?: string;
    desc?: string;
    make?: string;
    location?: string;
    drive?: string;
    transmission?: string;
    weight?: string;
    totalPrice?: string;
    perKm?: string;
    id?: string;
    power?: string;
}

export interface AddCarType {
    model?: string;
    desc?: string;
    make?: string;
    location?: string;
    drive?: string;
    transmission?: string;
    weight?: string;
    totalPrice?: string;
    perKm?: string;
    id?: string;
    power?: string;
}